+++ packages/@proto-graphql/protoc-plugin-helpers/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @proto-graphql/protoc-plugin-helpers
 
+## 0.2.0
+
+### Minor Changes
+
+- [#224](https://github.com/proto-graphql/proto-graphql-js/pull/224) [`6c25f40`](https://github.com/proto-graphql/proto-graphql-js/commit/6c25f4035f1b788a63bf005dbc52ca5d0bd5f2a2) Thanks [@izumin5210](https://github.com/izumin5210)! - support `long_number` param for mapping 64bit numbers to GraphQL Int
+
+### Patch Changes
+
+- Updated dependencies [[`6c25f40`](https://github.com/proto-graphql/proto-graphql-js/commit/6c25f4035f1b788a63bf005dbc52ca5d0bd5f2a2)]:
+  - @proto-graphql/codegen-core@0.2.0
+
+## 0.1.1
+
+### Patch Changes
+
+- [#213](https://github.com/proto-graphql/proto-graphql-js/pull/213) [`c5b988a`](https://github.com/proto-graphql/proto-graphql-js/commit/c5b988a4a1eac5d56d29572b593ae7643e23bd88) Thanks [@izumin5210](https://github.com/izumin5210)! - handle `pothos_builder_path` param correctly
+
+- Updated dependencies [[`ba647eb`](https://github.com/proto-graphql/proto-graphql-js/commit/ba647eb584850fee9c632a76cc1c028ce8ccd725)]:
+  - @proto-graphql/codegen-core@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes