+++ packages/protoc-gen-nexus/CHANGELOG.md
@@ -1,5 +1,32 @@
 # protoc-gen-nexus
 
+## 0.5.4
+
+### Patch Changes
+
+- Updated dependencies [[`6c25f40`](https://github.com/proto-graphql/proto-graphql-js/commit/6c25f4035f1b788a63bf005dbc52ca5d0bd5f2a2)]:
+  - @proto-graphql/codegen-core@0.2.0
+  - @proto-graphql/protoc-plugin-helpers@0.2.0
+
+## 0.5.3
+
+### Patch Changes
+
+- [#217](https://github.com/proto-graphql/proto-graphql-js/pull/217) [`5e2e85b`](https://github.com/proto-graphql/proto-graphql-js/commit/5e2e85baf6e0e21960104d3db30a4e21e04e1627) Thanks [@izumin5210](https://github.com/izumin5210)! - add `eslint-disable` directive to generated files
+
+- Updated dependencies [[`c1246c3`](https://github.com/proto-graphql/proto-graphql-js/commit/c1246c3a349f11e3b2bbfd6198c1a06b37270ece)]:
+  - @proto-graphql/codegen-core@0.1.2
+
+## 0.5.2
+
+### Patch Changes
+
+- [#212](https://github.com/proto-graphql/proto-graphql-js/pull/212) [`ba647eb`](https://github.com/proto-graphql/proto-graphql-js/commit/ba647eb584850fee9c632a76cc1c028ce8ccd725) Thanks [@izumin5210](https://github.com/izumin5210)! - refactor dependencies
+
+- Updated dependencies [[`ba647eb`](https://github.com/proto-graphql/proto-graphql-js/commit/ba647eb584850fee9c632a76cc1c028ce8ccd725), [`c5b988a`](https://github.com/proto-graphql/proto-graphql-js/commit/c5b988a4a1eac5d56d29572b593ae7643e23bd88)]:
+  - @proto-graphql/codegen-core@0.1.1
+  - @proto-graphql/protoc-plugin-helpers@0.1.1
+
 ## 0.5.1
 
 ### Patch Changes